[ADD] eatman: ajout du status dans le cycle de calcul du cout de revient + modification de la vue
