+++ eatman/__manifest__.py
@@ -25,6 +25,7 @@
     'data': [
         'security/ir.model.access.csv',
         'views/views.xml',
+        'views/foodcostview_wizard.xml',
         'views/templates.xml',
 
         'demo/demo.xml',
