+++ eatman/models/.ipynb_checkpoints/__init__-checkpoint.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import models
-from . import product
-from . import receipe
+from . import product, receipe, foodcostwizard
