+++ eatman/views/.ipynb_checkpoints/foodcostview_wizard-checkpoint.xml
@@ -1,0 +1,45 @@
+<odoo>
+    <data>
+        <act_window id="action_all_foodcostcalculation"
+            name="Calcul de cout de revient"
+            res_model="eatman.foodcostwizard"
+            view_mode="tree,form"
+            limit="80" />
+        
+        <menuitem id="menu_all_foodcost_calculation"
+          parent="eatman_top"
+          action="action_all_foodcostcalculation"
+          sequence="20"
+          groups="" />
+      
+        
+        <record id="form_foodcostWizard" model="ir.ui.view">
+            <field name="name">Calcul du coût de revient</field>
+            <field name="model">eatman.foodcostwizard</field>
+            <field name="arch" type="xml">
+                <form>
+                    <group>
+                        <field name ="name"/>
+                        <field name ="status" readonly="1"/>
+                    </group>
+                    <footer>
+                        <button string="Lancer le calcul" type="object" name="foodcost_total"/>
+                    <button string="Annuler" special="cancel" class ="btn_btn-secondary"/>
+                    </footer>
+                </form>
+            </field>
+        </record>
+        <record id="tree_foodcostWizard" model="ir.ui.view">
+            <field name="name">Calcul du coût de revient</field>
+            <field name="model">eatman.foodcostwizard</field>
+            <field name="type">tree</field>
+            <field name="arch" type="xml">
+                <tree>
+                    
+                        <field name ="name"/>
+                        <field name ="status"/>
+                </tree>
+            </field>
+        </record>
+    </data>
+</odoo>
