+++ eatman/views/views.xml
@@ -110,11 +110,7 @@
             res_model="product.template"
             view_mode="tree,form"
             limit="80" />
-    <act_window id="action_all_foodcostcalculation"
-            name="Calcul de cout de revient"
-            res_model="eatman.foodcostwizard"
-            view_mode="tree,form"
-            limit="80" />
+
       
     <menuitem id="eatman_top"
         name="Foodcost" />
@@ -135,48 +131,26 @@
           action="action_all_product"
           sequence="20"
           groups="" />
-        <menuitem id="menu_all_foodcost_calculation"
-          parent="eatman_top"
-          action="action_all_foodcostcalculation"
-          sequence="20"
-          groups="" />
-      
-    <record id="form_all_receipe" model="ir.ui.view">
-        <field name="name">Mes recettes</field>
-        <field name="model">eatman.receipe</field>
-        <field name="arch" type="xml">
-            <form>
-                <group>
-                    <field name ="product_cooked"/>   
-                    <field name ="receipe_quantity"/>
-                    <field name ="receipe_uom"/>
-                    <field name ="receipe_line_ids" string="Ligne de recette">
-                        <tree string="Ligne de recette">
-                            <field name ="product_ingredient"/>
-                            <field name ="ingredient_quantity"/>
-                            <field name ="ingredient_uom"/>
-                        </tree>
-                    </field>
-                </group>
-            </form>
-        </field>
-    </record>
-      
-     <record id="form_foodcostWizard" model="ir.ui.view">
-        <field name="name">Calcul du coût de revient</field>
-        <field name="model">eatman.foodcostwizard</field>
-        <field name="arch" type="xml">
-            <form>
-                <group>
-                    <field name ="name"/>
-                 
-                </group>
-                <footer>
-                    <button string="Lancer le calcul" type="object" name="foodcost_total"/>
-                    <button string="Annuler" special="cancel" class ="btn_btn-secondary"/>
-                </footer>
-            </form>
-        </field>
-    </record>
-  </data>
+
+        <record id="form_all_receipe" model="ir.ui.view">
+            <field name="name">Mes recettes</field>
+            <field name="model">eatman.receipe</field>
+            <field name="arch" type="xml">
+                <form>
+                    <group>
+                        <field name ="product_cooked"/>   
+                        <field name ="receipe_quantity"/>
+                        <field name ="receipe_uom"/>
+                        <field name ="receipe_line_ids" string="Ligne de recette">
+                            <tree string="Ligne de recette">
+                                <field name ="product_ingredient"/>
+                                <field name ="ingredient_quantity"/>
+                                <field name ="ingredient_uom"/>
+                            </tree>
+                        </field>
+                    </group>
+                </form>
+            </field>
+        </record>
+    </data>
 </odoo>
